+++ lib/io/html/src/window.dart
@@ -1,4 +1,5 @@
 import 'package:RAL/html_i.dart' as interface;
+import 'package:RAL/io/html/src/storage.dart';
 
 class Window implements interface.Window {
   static final Window _window = new Window._internal();
@@ -13,7 +14,11 @@ class Window implements interface.Window {
 
   Location get location => _location;
 
-  interface.Storage get localStorage => null;
+  interface.Storage get localStorage => new Storage();
+
+  interface.Storage get sessionStorage => new Storage();
+
+  void open(String url, String target, [String features]) => null;
 }
 
 class Location implements interface.Location{
@@ -21,4 +26,18 @@ class Location implements interface.Location{
   String get hash => '';
 
   String get host => 'localhost';
+
+  String get hostname => '';
+
+  String get href => '';
+
+  String get origin =>'';
+
+  String get pathname => '';
+
+  String get port => '';
+
+  String get protocol => '';
+
+  String get search => '';
 }